Bug #1285853 power: supply: sc27xx: Fix the possible overflow
[root cause  ] Fix the possible overflow
[changes     ] Fix the possible overflow
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Sharkle
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

When converting the ADC values to the current or voltage values,
the adc * 1000 may exceeds since adc is int type.

+++ drivers/power/supply/sc27xx_fuel_gauge.c
@@ -192,12 +192,12 @@ static const char * const sc27xx_charger_supply_name[] = {
 	"bq2560x_charger",
 };
 
-static int sc27xx_fgu_adc_to_current(struct sc27xx_fgu_data *data, int adc)
+static int sc27xx_fgu_adc_to_current(struct sc27xx_fgu_data *data, s64 adc)
 {
 	return DIV_ROUND_CLOSEST(adc * 1000, data->cur_1000ma_adc);
 }
 
-static int sc27xx_fgu_adc_to_voltage(struct sc27xx_fgu_data *data, int adc)
+static int sc27xx_fgu_adc_to_voltage(struct sc27xx_fgu_data *data, s64 adc)
 {
 	return DIV_ROUND_CLOSEST(adc * 1000, data->vol_1000mv_adc);
 }
